Add scores

+++ server/gateway/users/users.go
@@ -36,7 +36,7 @@ type NewUser struct {
 type Score struct {
 	ID      int64     `json:"-"`
 	Score   int       `json:"score"`
-	UserID  int64     `json:"userID"`
+	UserID  int       `json:"userID"`
 	Created time.Time `json:"created"`
 }
